+++ docs/changelog.md
@@ -24,6 +24,10 @@ a full list of changes to the project, please consult the commit log.
 - Ancient Slate:
     - Armor reduction base value is now `-4`.
     - Each additional Ancient Slate adds `-4` to the armor reduction.
+- Dark Emerald / Enchanted Emerald:
+    - Each additional Dark Emerald or Enchanted Emerald:
+        - Causes attacks to bounce to an additional nearby enemy.
+        - Increases the chance to stun on attack.
 - Emeralds:
     - Poison damage can now kill units.
 - Fire Star: